rename relocated_models directory, remove original models directory, add imports, resolve circular dependencies
